requirements.txt
imutils
numpy
Pillow
scipy
streamlit
tensorflow
opencv-python-headless